Torbray dev3 (#34)
* Tightening For Loop

Maybe the option was there to add extra bits for Rare or Unique? I left the if clause in, but could be removed if there's nothing outside of Normal, Rare or Unique

* Tightening 'show_inventory'

Thanks for the opportunity to learn GitHub! I'm figuring out what the branches are now :)

Starting to clean up print statements

* Update PlayerClass.py

* Removing Passes

Removed all passes in the code, and some funny true statements

* Separate monsters

Might work

* Indents

* Updating Player Stats

* Minor Update

* Fixes to Out of Bounds and Selling Spells

Co-authored-by: Arivald8 <59714150+Arivald8@users.noreply.github.com>